--- a/libr/anal/p/anal_pic.c
+++ b/libr/anal/p/anal_pic.c
@@ -1006,37 +1006,39 @@ beach:
 }
 
 static int anal_pic_midrange_set_reg_profile (RAnal *esil) {
-	const char *p;
-	p = "=PC	pc\n"
-	    "=SP	stkptr\n"
-	    "gpr	indf0	.8	0	0\n"
-	    "gpr	indf1	.8	1	0\n"
-	    "gpr	pcl		.8	2	0\n"
-	    "gpr	status	.8	3	0\n"
-	    "flg	c		.1	3.0	0\n"
-	    "flg	dc		.1	3.1	0\n"
-	    "flg	z		.1	3.2	0\n"
-	    "flg	pd		.1	3.3	0\n"
-	    "flg	to		.1	3.4	0\n"
-	    "gpr	fsr0l	.8	4	0\n"
-	    "gpr	fsr0h	.8	5	0\n"
-	    "gpr	fsr1l	.8	6	0\n"
-	    "gpr	fsr1h	.8	7	0\n"
-	    "gpr	bsr		.8	8	0\n"
-	    "gpr	wreg	.8	9	0\n"
-	    "gpr	pclath	.8	10	0\n"
-	    "gpr	intcon	.8	11	0\n"
-	    "gpr	pc		.16	12	0\n"
-	    "gpr	stkptr	.8	14	0\n"
-	    "gpr	_sram	.32 15	0\n"
-	    "gpr	_stack	.32 19	0\n";
+	RStrBuf sb;
+	const char *p = \
+		"=PC	pc\n"
+		"=SP	stkptr\n"
+		"gpr	indf0	.8	0	0\n"
+		"gpr	indf1	.8	1	0\n"
+		"gpr	pcl		.8	2	0\n"
+		"gpr	status	.8	3	0\n"
+		"flg	c		.1	3.0	0\n"
+		"flg	dc		.1	3.1	0\n"
+		"flg	z		.1	3.2	0\n"
+		"flg	pd		.1	3.3	0\n"
+		"flg	to		.1	3.4	0\n"
+		"gpr	fsr0l	.8	4	0\n"
+		"gpr	fsr0h	.8	5	0\n"
+		"gpr	fsr1l	.8	6	0\n"
+		"gpr	fsr1h	.8	7	0\n"
+		"gpr	bsr		.8	8	0\n"
+		"gpr	wreg	.8	9	0\n"
+		"gpr	pclath	.8	10	0\n"
+		"gpr	intcon	.8	11	0\n"
+		"gpr	pc		.16	12	0\n"
+		"gpr	stkptr	.8	14	0\n"
+		"gpr	_sram	.32 15	0\n"
+		"gpr	_stack	.32 19	0\n";
 
-	return r_reg_set_profile_string (esil->reg, p);
+	r_strbuf_init_const (&sb, p, strlen (p));
+	return r_reg_set_profile_string (esil->reg, &sb);
 }
 
 static int anal_pic_pic18_set_reg_profile(RAnal *esil) {
-	const char *p;
-	p =
+	RStrBuf sb;
+	const char *p = \
 		"#pc lives in nowhere actually"
 		"=PC	pc\n"
 		"=SP	tos\n"
@@ -1149,7 +1151,8 @@ static int anal_pic_pic18_set_reg_profile(RAnal *esil) {
 		"gpr	stkptr	.8	96	0\n"
 		"gpr	tablat	.8	14	0\n";
 
-	return r_reg_set_profile_string (esil->reg, p);
+	r_strbuf_init_const (&sb, p, strlen (p));
+	return r_reg_set_profile_string (esil->reg, &sb);
 }