--- a/libr/anal/p/anal_riscv.c
+++ b/libr/anal/p/anal_riscv.c
@@ -546,7 +546,8 @@ static int riscv_op(RAnal *anal, RAnalOp *op, ut64 addr, const ut8 *data, int le
 	return op->size;
 }
 
-static char *get_reg_profile(RAnal *anal) {
+static RStrBuf *get_reg_profile(RAnal *anal) {
+	int l = 0;
 	const char *p = NULL;
 	switch (anal->bits) {
 	case 32: p =
@@ -633,6 +634,7 @@ static char *get_reg_profile(RAnal *anal) {
 		"flg	frm	.3	3077	0\n"
 		;
 
+		l = strlen (p);
 		break;
 	case 64: p =
 		"=PC	pc\n"
@@ -717,9 +719,10 @@ static char *get_reg_profile(RAnal *anal) {
 		"flg	frm	.3	4101	0\n"
 		;
 
+		l = strlen (p);
 		break;
 	}
-	return (p && *p)? strdup (p): NULL;
+	return (p && *p)? r_strbuf_new_const (p, l): NULL;
 }
 
 RAnalPlugin r_anal_plugin_riscv = {