--- a/libr/anal/p/anal_sparc_cs.c
+++ b/libr/anal/p/anal_sparc_cs.c
@@ -327,6 +327,7 @@ static int analop(RAnal *a, RAnalOp *op, ut64 addr, const ut8 *buf, int len, RAn
 }
 
 static int set_reg_profile(RAnal *anal) {
+	RStrBuf sb;
 	const char *p = \
 		"=PC	pc\n"
 		"=SP	y\n"
@@ -375,7 +376,8 @@ static int set_reg_profile(RAnal *anal) {
 		"gpr	r30	.32	136	0\n"
 		"gpr	r31	.32	140	0\n"
 	;
-	return r_reg_set_profile_string (anal->reg, p);
+	r_strbuf_init_const (&sb, p, strlen (p));
+	return r_reg_set_profile_string (anal->reg, &sb);
 }
 
 static int archinfo(RAnal *anal, int q) {