--- a/libr/anal/p/anal_v810.c
+++ b/libr/anal/p/anal_v810.c
@@ -392,6 +392,7 @@ static int v810_op(RAnal *anal, RAnalOp *op, ut64 addr, const ut8 *buf, int len,
 }
 
 static int set_reg_profile(RAnal *anal) {
+	RStrBuf sb;
 	const char *p =
 		"=PC	pc\n"
 		"=SP	r3\n"
@@ -444,7 +445,8 @@ static int set_reg_profile(RAnal *anal) {
 		"flg	s   .1 132.30 0\n"
 		"flg	z   .1 132.31 0\n";
 
-	return r_reg_set_profile_string (anal->reg, p);
+	r_strbuf_init_const (&sb, p, strlen (p));
+	return r_reg_set_profile_string (anal->reg, &sb);
 }
 
 RAnalPlugin r_anal_plugin_v810 = {