… libr/anal/p/anal_x86_cs.c
… libr/core/cmd_debug.c
… libr/debug/p/debug_bf.c
… libr/debug/p/debug_bochs.c
… libr/debug/p/debug_esil.c
… libr/debug/p/debug_gdb.c
… libr/debug/p/debug_io.c
… libr/debug/p/debug_native.c
… libr/debug/p/debug_qnx.c
… libr/debug/p/debug_rap.c
… libr/debug/p/native/linux/linux_debug.c
… libr/debug/p/native/linux/linux_debug.h
… libr/debug/p/native/linux/reg/linux-arm.h
… libr/debug/p/native/linux/reg/linux-arm64.h
… libr/debug/p/native/linux/reg/linux-mips.h
… libr/debug/p/native/linux/reg/linux-mips64.h
… libr/debug/p/native/linux/reg/linux-ppc.h
… libr/debug/p/native/linux/reg/linux-ppc64.h
… libr/debug/p/native/linux/reg/linux-riscv64.h
… libr/debug/p/native/linux/reg/linux-x64-32.h
… libr/debug/p/native/linux/reg/linux-x64.h
… libr/debug/p/native/linux/reg/linux-x86.h
… libr/debug/p/native/reg.c
… libr/debug/p/native/reg/kfbsd-arm64.h
… libr/debug/p/native/reg/kfbsd-x64.h
… libr/debug/p/native/reg/kfbsd-x86.h
… libr/debug/p/native/reg/netbsd-x64.h
… libr/debug/p/native/reg/netbsd-x86.h
… libr/debug/p/native/reg/windows-x64.h
… libr/debug/p/native/reg/windows-x86.h