In the most cases string passed to `r_reg_set_profile_string()` is a string literal. And some of them are 3k characters long. We can use RStrBuf objects for storing pointer to a const string which will allow us to avoid:
1) Calling `strdup()` in `r_anal_get_reg_profile()` (just alloc a new RStrBuf object)
2) Calling `strdup()` in `r_reg_set_profile_string()` for caching reg profile string (just copy the RStrBuf object)
3) Calling `strcmp()` in `r_reg_set_profile_string()` for comparing reg profile string (just compare pointers).

But these changes gives tangible analysis speedup only for multibit binaries (e.g. ARM/Thumb) and only on Windows.

### Windows (VS 2017/release/shared)
`Measure-Command {.radare2.exe -Q -qq -A libcrackme.so}`
#### Before:
```
Minutes           : 0
Seconds           : 55
...
```
#### After:
```
Minutes           : 0
Seconds           : 49
...
```
### GNU/Linux (gcc 8.3.0/release/shared)
`time radare2 -A -qq -Q libcrackme.so`
#### Before:
```
real	0m37,300s
user	0m36,576s
sys	0m0,702s
```
#### After:
```
real	0m35,684s
user	0m34,831s
sys	0m0,820s
```