@ret2libc 2019-11-06
Isn't r_str_constpool supposed to do just that?