@ret2libc 2019-11-06
> > Isn't r_str_constpool supposed to do just that?
> 
> I think using `r_str_constpool` will make analysis much slower than now.

Mh, after looking at the implementation (which uses sdb_ht, when it should be using HtPP), I kinda agree...