@ret2libc 2019-11-06
But still, I don't like too much making RStrBuf even more complicated to support this const case, which should be the perfect usecase for something like r_strconstpool (which may need to be improved). If these strings are constant and don't need to be changed, why do we strdup them all around?