@pelijah 2019-11-06
Profile can be loaded from a file:
https://github.com/radareorg/radare2/blob/36e557aa9619c1900e0fe62d9e7048475528ca1c/libr/reg/profile.c#L238

Or from debug plugins:
https://github.com/radareorg/radare2/blob/571a080ca732ba585686a5580de2d14f18453a29/libr/debug/p/debug_io.c#L77
https://github.com/radareorg/radare2/blob/20e19d602337d29d9e2831ddf3f8b46168b9450a/libr/debug/p/debug_rap.c#L51
https://github.com/radareorg/radare2/blob/82b02e8b89aad92d879e376e519b10edd715ff55/libr/debug/p/debug_gdb.c#L454