@pelijah 2019-11-09
Therefore you can't just remove `strdup()`.
I don't think that my changes make RStrBuf code super complicated. Anyway we need to make similar code changes in `strbuf.c` if  we want to add ref counting feature in the future.