@ret2libc 2019-11-11
@pelijah did you have a look at r_strconstpool? it should give you exactly what you need with regard to this const thing. @thestr4ng3r have improved it a bit, so it should be fast enough IMHO. You should just use `r_str_constpool_get` and then you can compare pointers once you are sure you are dealing with constpool strings.