@pelijah 2019-11-11
Using `r_str_constpool_get()` gives almost no effect to analysis speed (compared to current implementation with `strdup() + strcmp() + strdup()`).