@pelijah 2019-11-12
During analysis of some binaries r2 modifies `asm.bits` (e.g. on ARM binaries it often switches between 16/32 bits) that leads to a call of `r_anal_set_reg_profile()` function. This function can implicitly invoke `r_reg_set_profile_string()` via `set_reg_profile` method of the current anal plugin, or explicitly pass  to `r_reg_set_profile_string()` a reg profile string obtained from `get_reg_profile` method. `r_reg_set_profile_string()` uses `strdup()` to cache the reg profile string and `strcmp()` to avoid parsing unchanged reg profiles.

In sum: `r_anal_set_reg_profile()` is called too many times -> `strcmp()` + `strdup()` are called too many times. Don't forget that some reg profile strings are 3k characters long. Also `get_reg_profile` method returns a pointer to a heap allocated string so don't forget about extra `strdup()`.

In the most cases a string passed to `r_reg_set_profile_string()` is a constant string. No need to spend CPU time on these `strcmp()` and `strdup()`. We can just compare pointers and save it directly. But in some cases we have to pass a heap allocated string to `r_reg_set_profile_string()`. So sometimes we stiil need to use `strcmp()` and `strdup()`. To resolve this issue we need to have some meta about the pointer to a string passed to `r_reg_set_profile_string()`. We can add a new boolean argument for `r_reg_set_profile_string()` or use some container. I think RStrBuf object is a quite suitable container for this.

I have to say I understand that one use case of the presented feature is not enough to add it.