@ret2libc 2019-11-12
> Using `r_str_constpool_get()` gives almost no effect to analysis speed (compared to current implementation with `strdup() + strcmp() + strdup()`).
> Current:
> ![изображение](https://user-images.githubusercontent.com/16493991/68626481-83e77b00-04ec-11ea-951c-8bc56de0be7f.png)
> Constpool (here almost all CPU time of `r_reg_set_profile_string()` is spent on `r_str_constpool_get()`):
> ![изображение](https://user-images.githubusercontent.com/16493991/68626549-a24d7680-04ec-11ea-94b1-0b36022f90fe.png)

Did you do the comparisons with pointers after you are sure the strings are coming from constpool? Or you are just calling r_strconstpool_get again and again?