@radare 2019-11-13
ah wait, cant be const * because we need to construct them from other langs too

> On 12 Nov 2019, at 23:24, Paul I. <notifications@github.com> wrote:
> 
> During analysis of some binaries r2 modifies asm.bits (e.g. on ARM binaries it often switches between 16/32 bits) that leads to a call of r_anal_set_reg_profile() function. This function can implicitly invoke r_reg_set_profile_string() via set_reg_profile method of the current anal plugin, or explicitly pass to r_reg_set_profile_string() a reg profile string obtained from get_reg_profile method. r_reg_set_profile_string() uses strdup() to cache the reg profile string and strcmp() to avoid parsing unchanged reg profiles.
> 
> In sum: r_anal_set_reg_profile() is called too many times -> strcmp() + strdup() are called too many times. Don't forget that some reg profile strings are 3k characters long. Also get_reg_profile method returns a pointer to a heap allocated string so don't forget about extra strdup().
> 
> In the most cases a string passed to r_reg_set_profile_string() is a constant string. No need to spend CPU time on these strcmp() and strdup(). We can just compare pointers and save it directly. But in some cases we have to pass a heap allocated string to r_reg_set_profile_string(). So sometimes we stiil need to use strcmp() and strdup(). To resolve this issue we need to have some meta about the pointer to a string passed to r_reg_set_profile_string(). We can add a new boolean argument for r_reg_set_profile_string() or use some container. I think RStrBuf object is a quite suitable container for this.
> 
> I have to say I understand that one use case of the presented feature is not enough to add it.
> 
> —
> You are receiving this because you commented.
> Reply to this email directly, view it on GitHub <https://github.com/radareorg/radare2/pull/15400?email_source=notifications&email_token=AAG75FRUJKDW6HY2CJXSTIDQTMUJ5A5CNFSM4JIAXSQ2YY3PNVWWK3TUL52HS4DFVREXG43VMVBW63LNMVXHJKTDN5WW2ZLOORPWSZGOED4FFOA#issuecomment-553145016>, or unsubscribe <https://github.com/notifications/unsubscribe-auth/AAG75FWRYQ5AHOVXL7RA7ETQTMUJ5ANCNFSM4JIAXSQQ>.
>