@radare 2019-11-14
just read the code:

typedef struct {
        char buf[32];
        int len;
        char *ptr;
        int ptrlen;
} RStrBuf;

ptr=buf unless string length > 32

> On 13 Nov 2019, at 22:14, Paul I. <notifications@github.com> wrote:
> 
> @pelijah commented on this pull request.
> 
> In libr/anal/anal.c <https://github.com/radareorg/radare2/pull/15400#discussion_r346004928>:
> 
> > @@ -265,7 +265,7 @@ R_API bool r_anal_use(RAnal *anal, const char *name) {
>  	return false;
>  }
>  
> -R_API char *r_anal_get_reg_profile(RAnal *anal) {
> +R_API RStrBuf *r_anal_get_reg_profile(RAnal *anal) {
> strbuf changes the pointer depending of the length of the string
> Can you explain more?
> 
> —
> You are receiving this because you commented.
> Reply to this email directly, view it on GitHub <https://github.com/radareorg/radare2/pull/15400?email_source=notifications&email_token=AAG75FRKQH7WMZJT74KRR7DQTRU5HA5CNFSM4JIAXSQ2YY3PNVWWK3TUL52HS4DFWFIHK3DMKJSXC5LFON2FEZLWNFSXPKTDN5WW2ZLOORPWSZGOCLPFJWQ#discussion_r346004928>, or unsubscribe <https://github.com/notifications/unsubscribe-auth/AAG75FRKK6PMKTTG3VHHW33QTRU5HANCNFSM4JIAXSQQ>.
>